chores: rename dicrectory
