+++ src/main/java/view/gui_components/OrderEntryPanel.java
@@ -1,4 +1,4 @@
-package view.GUI;
+package view.gui_components;
 
 import javax.swing.*;
 import java.awt.*;
