+++ src/main/java/view/gui_components/TradeSimulationFrame.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="view.GUI.TradeSimulationFrame">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="view.gui_components.TradeSimulationFrame">
   <grid id="27dc6" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="1" left="1" bottom="1" right="1"/>
     <constraints>
